requirements.txt
# Core dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Testing and evaluation
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Data handling and storage
datasets>=2.14.0
huggingface-hub>=0.17.0

# Logging and monitoring
wandb>=0.15.0
loguru>=0.7.0

# Development tools
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0 